docker/route-vehicles-retriever/requirements.txt
cassandra-driver
cassandra-sigv4
boto3